Add roles to the used Asciidoc theme
The default RISC-V theme is extended by creating a new theme that
creates semantic roles, allowing the definition of new keywords (roles)
that modify the text when used.
Currently two roles are defined:
- header style for examples as used in the Caches chapter to avoid them
  being too small
- format of the extensions, so all extensions use the same format
  (currently italic)

These roles have been applied to the example sections in the caches
chapter for the first role, and to an extension used in the PMCs
chapter (just for testing purposes, needs to be applied every time an
extension is used in the text).