Add support for specific formating (#17)
* Add roles to the used Asciidoc theme

The default RISC-V theme is extended by creating a new theme that
creates semantic roles, allowing the definition of new keywords (roles)
that modify the text when used.
Currently two roles are defined:
- header style for examples as used in the Caches chapter to avoid them
  being too small
- format of the extensions, so all extensions use the same format
  (currently italic)

These roles have been applied to the example sections in the caches
chapter for the first role, and to an extension used in the PMCs
chapter (just for testing purposes, needs to be applied every time an
extension is used in the text).

* Use extension role on occurrences of extension in the cache chapter

* Use extension role on occurrences of extension in the PMCs chapter

* Fix: add new line at the end of the file

Signed-off-by: Daniel Gracia Pérez <50202128+dgptha@users.noreply.github.com>

* Fix: end of line detected by the pre-commit script

Signed-off-by: Daniel Gracia Pérez <50202128+dgptha@users.noreply.github.com>

---------

Signed-off-by: Daniel Gracia Pérez <50202128+dgptha@users.noreply.github.com>