Adapt Makefile and root document to create document releases (#24)
When running make without the REVMARK flag the document will be created
in draft mode, where a default release text will be used and a DRAFT
image will be put in the background of every page.

For creating a release of the document run make with REVMARK set up.
Typically this will be done when the document version number will be
also defined.
For example:
```
VERSION="v1.0.2" REVMARK="Release\ (update)." make
```
Make sure to escape spaces, like in the example above.

Signed-off-by: Daniel Gracia Pérez <50202128+dgptha@users.noreply.github.com>